producer/requirements.txt
kafka-python==2.2.15
websocket-client==1.9.0
